requirements.txt
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
airportsdata==20250811
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anthropic==0.64.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.5
blis==1.3.0
blobfile==3.0.0
boto3==1.40.21
botocore==1.40.21
Brotli==1.1.0
build==1.3.0
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
cloudpickle==3.1.1
coloredlogs==15.0.1
colorlog==6.9.0
compressed-tensors==0.11.0
confection==0.1.5
contourpy==1.3.3
cryptography==45.0.6
cuda-bindings==12.9.2
cuda-pathfinder==1.2.0
cuda-python==12.9.0
cycler==0.12.1
cymem==2.0.11
datasets==4.0.0
decorator==5.2.1
decord==0.6.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
doclayout_yolo==0.0.4
einops==0.8.1
executing==2.2.0
faiss-cpu==1.12.0
faiss-gpu-cu12==1.12.0
fast-langdetect==0.2.5
fastapi==0.116.1
fasttext-predict==0.9.2.4
ffmpy==0.6.1
filelock==3.19.1
flashinfer-python==0.2.14.post1
flatbuffers==25.2.10
fonttools==4.59.2
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
gradio==5.44.1
gradio_client==1.12.1
gradio_pdf==0.0.22
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.9
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
interegular==0.3.3
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
json_repair==0.50.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
lazy_loader==0.4
litellm==1.76.0
llguidance==0.7.30
loguru==0.7.3
lxml==6.0.1
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mineru @ git+https://github.com/myhloli/Magic-PDF.git@bae254fa72a4d66473d6617d8a172172771af3a2
modelscope==1.29.1
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
murmurhash==1.0.13
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.14.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.24
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
onnxruntime==1.22.1
openai==1.99.1
openai-harmony==0.0.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.11.3
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.3.2
parso==0.8.5
partial-json-parser==0.2.1.1.post6
pdfminer.six==20250506
pdftext==0.6.3
pexpect==4.9.0
pillow==11.3.0
polars==1.32.3
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pybase64==1.4.2
pyclipper==1.3.0.post6
pycountry==24.6.1
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pynvml==12.0.0
pyparsing==3.2.3
pypdf==6.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
referencing==0.36.2
regex==2025.8.29
reportlab==4.4.3
requests==2.32.5
rich==14.1.0
robust-downloader==0.0.2
rpds-py==0.27.1
ruff==0.12.11
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.1
setproctitle==1.3.6
sgl-kernel==0.3.7
sglang==0.5.1.post3
shapely==2.1.1
shellingham==1.5.4
simsimd==6.5.1
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.3
stringzilla==3.12.6
sympy==1.14.0
thinc==8.3.6
thop==0.1.1.post2209072238
tifffile==2025.8.28
tiktoken==0.11.0
timm==1.0.16
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.8.0
torch_memory_saver==0.0.8
torchao==0.9.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.2
triton==3.4.0
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.189
ultralytics-thop==2.0.16
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.3
xgrammar==0.1.23
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
